pyproject.toml
[project]
name = "OEcon"
version = "0.2.0-dev"
description = "Python tools to convert and preprocess OpenEphys data to the DAQ-HDF5 (dh5) format"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dh5io",
    "open-ephys-python-tools",
    "scipy>=1.15.2",
    "numpy>2.0",
    "vstim-python-tools",
    "psutil>=7.2.2",
    "pydantic>=2.0",
    "pyside6>=6.7",
    "tqdm>=4.67.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.uv.sources]
vstim-python-tools = { git = "https://github.com/brain-bremen/vstim-python-tools.git" }


[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pytest>=8.3.5",
    "pytest-plt>=1.1.1",
    "scipy-stubs>=1.15.3.0",
    "types-tqdm>=4.67.3.20260303",
]


[project.scripts]
oecon = "cli.main:main"
oecon-gui = "gui.app:main"
oecon-inspect = "gui.inspector_app:main"
